+++ src/Camera.h
@@ -18,8 +18,8 @@ public:
 	Vector3 hinc;
 	Vector3 topLeft;
 	int h, v;
-	Camera(Vector3 position, Vector3 u, Vector3 v, Vector3 w, float width,
-			float height, float dist, int vresolution, int hresolution);
+	Camera(Vector3 position, Vector3 u, Vector3 v, Vector3 w, double width,
+			double height, double dist, int vresolution, int hresolution);
 	Ray makeRay(int h, int v);
 	virtual ~Camera();
 };
