+++ src/Matrix3.h
@@ -12,14 +12,14 @@
 #include <string>
 class Matrix3 {
 public:
-	float a, b, c, d, e, f, g, h, i;
-	Matrix3(float a, float b, float c, float d, float e, float f, float g, float h, float i);
+	double a, b, c, d, e, f, g, h, i;
+	Matrix3(double a, double b, double c, double d, double e, double f, double g, double h, double i);
 	Matrix3(Vector3 v1, Vector3 v2, Vector3 v3);
 	Matrix3 mult(Matrix3 other);
 	Vector3 trans(Vector3 v);
 	std::string repr();
 	virtual ~Matrix3();
 };
-Matrix3 horizMat(float angle);
-Matrix3 vertMat(float angle);
+Matrix3 horizMat(double angle);
+Matrix3 vertMat(double angle);
 #endif /* MATRIX3_H_ */
