+++ src/Plane.h
@@ -19,7 +19,7 @@ public:
 	Vector3 point, norm;
 	Color c;
 	Plane(Vector3 point, Vector3 normal, PhongProfile prof);
-	virtual float intersect(Ray other);
+	virtual double intersect(Ray other);
 	virtual Vector3 normal(Vector3 place);
 	virtual ~Plane();
 };
