+++ src/Ray.cpp
@@ -20,7 +20,7 @@ std::string Ray::repr(){
 	return "{\n start:" + start.repr() + "\n dir:" + dir.repr() + "\n}";
 }
 
-Vector3 Ray::getPoint(float t){
+Vector3 Ray::getPoint(double t){
 	if(t < 0){
 		throw 69;
 	}
